+++ Lodjinha/Lodjinha/UseCases/Entities/Banner.swift
@@ -1,0 +1,13 @@
+//
+//  Banner.swift
+//  Lodjinha
+//
+//  Created by Douglas da Silva Santos on 27/03/19.
+//  Copyright © 2019 Douglas da Silva Santos. All rights reserved.
+//
+
+import Foundation
+
+struct Banner {
+    let imageUrl: String
+}
